# docs/site/public/openapi.json
{
  "components": {
    "schemas": {
...
      "DeploySiteResponse": {
        "properties": {
          "deployment_number": {
            "title": "Deployment Number",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "private": {
            "title": "Private",
            "type": "boolean"
          },
          "site_name": {
            "title": "Site Name",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "site_name",
          "url",
          "private",
          "deployment_number"
        ],
        "title": "DeploySiteResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SiteDeploymentResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "credential": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential"
          },
          "deployed_at": {
            "format": "date-time",
            "title": "Deployed At",
            "type": "string"
          },
          "deployment_number": {
            "title": "Deployment Number",
            "type": "integer"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "source": {
            "enum": [
              "dashboard",
              "api"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "deployment_number",
          "deployed_at",
          "size_bytes",
          "source",
          "actor",
          "credential",
          "active"
        ],
        "title": "SiteDeploymentResponse",
        "type": "object"
      },
...
      "SiteResponse": {
        "properties": {
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "last_deployed_at": {
            "format": "date-time",
            "title": "Last Deployed At",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "private": {
            "title": "Private",
            "type": "boolean"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "total_views": {
            "title": "Total Views",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "created",
          "last_deployed_at",
          "size_bytes",
          "total_views",
          "private"
        ],
        "title": "SiteResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "A Buzz session or deployment token. Most operations require a session. Deployment tokens are accepted only for deployment to their assigned site.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP API for deploying and managing sites on a self-hosted Buzz server. API operations are available only on the configured Buzz domain.",
    "title": "Buzz",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/deploy": {
      "post": {
        "description": "Upload a ZIP archive as a new deployment. It becomes active immediately. Buzz creates the site if needed. A deployment token may deploy only to its assigned site and must send that name in X-Buzz-Site.",
        "operationId": "deploySite",
        "parameters": [
          {
            "description": "Site to deploy to. Buzz generates a name when omitted.",
            "in": "header",
            "name": "x-buzz-site",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Site to deploy to. Buzz generates a name when omitted.",
              "title": "X-Buzz-Site"
            }
          },
          {
            "description": "Set to 'private' to publish the site and protect it atomically.",
            "in": "header",
            "name": "x-buzz-access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Set to 'private' to publish the site and protect it atomically.",
              "title": "X-Buzz-Access"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "description": "A ZIP archive containing the site's files.",
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploySiteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The site name or archive is invalid."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The credential cannot deploy this site."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A deployment limit was exceeded."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Deploy a site",
        "tags": [
          "Sites"
        ]
      }
    },
...
    "/sites": {
      "get": {
        "operationId": "listSites",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SiteResponse"
                  },
                  "title": "Response Listsites",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A session token is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List owned sites",
        "tags": [
          "Sites"
        ]
      }
    },
    "/sites/{name}": {
      "delete": {
        "operationId": "deleteSite",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A session token and site ownership are required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The site does not exist."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Every custom domain must complete removal before deleting the site."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete a site",
        "tags": [
          "Sites"
        ]
      }
    },
    "/sites/{name}/deployments": {
      "get": {
        "operationId": "listSiteDeployments",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SiteDeploymentResponse"
                  },
                  "title": "Response Listsitedeployments",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List site deployments",
        "tags": [
          "Sites"
        ]
      }
    },
    "/sites/{name}/deployments/{deployment_number}/activate": {
      "post": {
        "operationId": "activateSiteDeployment",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "deployment_number",
            "required": true,
            "schema": {
              "title": "Deployment Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteDeploymentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Activate a site deployment",
        "tags": [
          "Sites"
        ]
      }
    },
...
  },
...
}